Rename transaction module to tx
