+++ crates/sovereign/sovereign-rollup-components/src/types/mod.rs
@@ -7,4 +7,4 @@ pub mod messages;
 pub mod payloads;
 pub mod rollup_id;
 pub mod status;
-pub mod transaction;
+pub mod tx;
